Added class_level_category_id field to progress
A ClassProgress should be tied to a student and a class. It should be
tied to a student and the class level category. I added the
class_level_category_id field to the ClassProgress schema definition
and created the associated aggregate and composite relationships. I
also removed the relationship between ClassProgress and Class but
didn’t remove the class_id field yet from the progress definition.