Redesigned Certificate Module
Redesigned the certificate module so that only one template of each
type could be uploaded and when you updated a template it updates it
for all certificates that use that template. Essentially, the
certificate is no longer tied to the primary key of the certificate
template. It’s all based on the certificate type.