Added checks to enrolment process
Now when a student is enrolled, the system checks to see if that
student is already enrolled in that scheduled class. If the student
isn’t enrolled in the scheduled class, then the system checks to see if
the ClassProgress object exists already for that level category. I
believe this will fixes #53.