Major update to isAnniversaryMilestone function
Changed the isAnniversaryMilestone function so that when given a date,
it returns the anniversary of the student if their anniversary date is
within a given day range from the date provided. The day range is a
system setting called “studentcentre.ann_day_range”. If no day range is
specified, the default is 30. The function makes use of the DateTime
class to add and subtract dates.